Removed and changed large chunks of the code to remove the XML 'middleman'. Committing just in case I mess up. (I think I already did). Also added a to readable string for the universe for debugging. Will be SUPER helpful
For assets/scripts/universeGen/main.py, I removed the xml generation part. Removed the parse method for ConquerSpace.game.universe.Universe, because it is a counterpart for the python script. But, it may come in handy in the future, so need to remember this commit.

The toReadableString() methods of all the ConquerSpace.game.universe turn the universe into a readable string. This has not been tested, though.

Changed the enums ConquerSpace.game.universe.StarTypes and ConquerSpace.game.universe.PlanetTypes to classes, because python does not have enums.

NOTE: need to add a readme.